* Resettle the RestAPI related services and fix multiple bugs
