+++ src/main/java/com/webank/weid/rpc/CptService.java
@@ -50,14 +50,6 @@ public interface CptService {
     ResponseData<CptBaseInfo> registerCpt(CptMapArgs args, Integer cptId);
 
     /**
-     * Register a new CPT to the blockchain by sending preset transaction hex value.
-     *
-     * @param transactionHex the transaction hex value
-     * @return The registered CPT info
-     */
-    ResponseData<String> registerCpt(String transactionHex);
-
-    /**
      * Register a new CPT to the blockchain.
      *
      * @param args the args
