+++ src/test/java/com/webank/weid/full/weid/TestCreateWeId1.java
@@ -42,6 +42,8 @@ import com.webank.weid.exception.WeIdBaseException;
 import com.webank.weid.full.TestBaseServcie;
 import com.webank.weid.protocol.response.CreateWeIdDataResult;
 import com.webank.weid.protocol.response.ResponseData;
+import com.webank.weid.rpc.RawTransactionService;
+import com.webank.weid.service.impl.RawTransactionServiceImpl;
 
 /**
  * non parametric createWeId method for testing WeIdService.
@@ -197,7 +199,8 @@ public class TestCreateWeId1 extends TestBaseServcie {
     @Test
     public void testCreateWeIdCase7() {
         String hex = StringUtils.EMPTY;
-        ResponseData<String> response = weIdService.createWeId(hex);
+        RawTransactionService rawTransactionService = new RawTransactionServiceImpl();
+        ResponseData<String> response = rawTransactionService.createWeId(hex);
         Assert.assertEquals(ErrorCode.ILLEGAL_INPUT.getCode(),
             response.getErrorCode().intValue());
         Assert.assertTrue(StringUtils.isEmpty(response.getResult()));
@@ -209,7 +212,8 @@ public class TestCreateWeId1 extends TestBaseServcie {
     @Test
     public void testCreateWeIdCase8() {
         String hex = "11111";
-        ResponseData<String> response = weIdService.createWeId(hex);
+        RawTransactionService rawTransactionService = new RawTransactionServiceImpl();
+        ResponseData<String> response = rawTransactionService.createWeId(hex);
         Assert.assertEquals(ErrorCode.TRANSACTION_EXECUTE_ERROR.getCode(),
             response.getErrorCode().intValue());
         Assert.assertTrue(StringUtils.isEmpty(response.getResult()));
